+++ Connectors/SpecFlow.VisualStudio.SpecFlowConnector.Generic/SpecFlowProxies/BindingRegistryFactoryBeforeV310000.cs
@@ -1,0 +1,15 @@
+﻿using TechTalk.SpecFlow;
+
+namespace SpecFlowConnector.SpecFlowProxies;
+
+public class BindingRegistryFactoryBeforeV310000 : BindingRegistryFactoryVLatest
+{
+    public BindingRegistryFactoryBeforeV310000(ILogger log) : base(log)
+    {
+    }
+
+    protected override ITestRunner InvokeCreateTestRunner(TestRunnerManager testRunnerManager)
+    {
+        return testRunnerManager.ReflectionCallMethod<ITestRunner>(nameof(TestRunnerManager.CreateTestRunner), 0);
+    }
+}